requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
google-adk
google-genai
httpx
deprecated
aiofiles
